requirements.txt
strands-agents
ddgs
fastapi
uvicorn[standard]
asyncio
rich
pytest
httpx
python-dotenv
